ParishSoftv2: separate active from deceased when searching for Members
When we load PS Members, the default is to not load inactive members
(inactive also includes deceased).  However, if we *do* load inactive
members, that *also* included loading deceased members -- which is not
(usually?) what we want.  So separate those two things: the desire to
load inactive members and the desire to load deceased members.

Meaning: if you ask to load inactive members, by default, we'll still
filter out deceased members.

Signed-off-by: Jeff Squyres <jeff@squyres.com>